requirements.txt
# Core PyTorch and ML libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Image processing and computer vision
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0

# Data science and numerical computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization and plotting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# System monitoring and utilities
psutil>=5.9.0
tqdm>=4.65.0

# Web interface (for tablet canvas integration)
flask>=2.3.0
flask-cors>=4.0.0

# Model export and optimization
onnx>=1.14.0
onnxruntime>=1.15.0

# Development and debugging
jupyter>=1.0.0
ipywidgets>=8.0.0

# Optional: For advanced augmentations
albumentations>=1.3.0

# Optional: For model analysis
tensorboard>=2.13.0

